requirements.txt
streamlit
python-docx
pdfplumber
pdf2image
pytesseract
huggingface_hub
python-dotenv
Pillow
